requirements.txt
# OpenMOVR App - Dependencies
# For Streamlit Cloud deployment

# Core data science
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
scipy>=1.10.0

# Configuration
pyyaml>=6.0

# Web framework
streamlit>=1.28.0
plotly>=5.17.0

# Excel support (for exports)
openpyxl>=3.1.0
xlsxwriter>=3.1.0
